refactor, fix flake8/mypy
