+++ core/dbt/adapters/base/__init__.py
@@ -4,7 +4,11 @@
 from dbt.contracts.connection import Credentials  # noqa
 from dbt.adapters.base.meta import available  # noqa
 from dbt.adapters.base.connections import BaseConnectionManager  # noqa
-from dbt.adapters.base.relation import BaseRelation, RelationType  # noqa
+from dbt.adapters.base.relation import (  # noqa
+    BaseRelation,
+    RelationType,
+    SchemaSearchMap,
+)
 from dbt.adapters.base.column import Column  # noqa
 from dbt.adapters.base.impl import BaseAdapter  # noqa
 from dbt.adapters.base.plugin import AdapterPlugin  # noqa
