+++ core/dbt/adapters/base/impl.py
@@ -31,7 +31,7 @@ from dbt.utils import filter_null_values, executor
 from dbt.adapters.base.connections import BaseConnectionManager, Connection
 from dbt.adapters.base.meta import AdapterMeta, available
 from dbt.adapters.base.relation import (
-    ComponentName, BaseRelation, InformationSchema
+    ComponentName, BaseRelation, InformationSchema, SchemaSearchMap
 )
 from dbt.adapters.base import Column as BaseColumn
 from dbt.adapters.cache import RelationsCache
@@ -118,62 +118,6 @@ def _relation_name(rel: Optional[BaseRelation]) -> str:
         return 'null relation'
     else:
         return str(rel)
-
-
-class SchemaSearchMap(Dict[InformationSchema, Set[Optional[str]]]):
-    """A utility class to keep track of what information_schema tables to
-    search for what schemas
-    """
-    def add(self, relation: BaseRelation, preserve_case=False):
-        key = relation.information_schema_only()
-        if key not in self:
-            self[key] = set()
-        schema: Optional[str] = None
-        if relation.schema is not None:
-            if preserve_case:
-                schema = relation.schema
-            else:
-                schema = relation.schema.lower()
-        self[key].add(schema)
-
-    def search(self) -> Iterator[Tuple[InformationSchema, Optional[str]]]:
-        for information_schema_name, schemas in self.items():
-            for schema in schemas:
-                yield information_schema_name, schema
-
-    def schemas_searched(self) -> Set[Tuple[str, Optional[str]]]:
-        result: Set[Tuple[str, Optional[str]]] = set()
-        for information_schema_name, schemas in self.items():
-            if information_schema_name.database is None:
-                raise InternalException(
-                    'Got a None database in an information schema!'
-                )
-            result.update(
-                (information_schema_name.database, schema)
-                for schema in schemas
-            )
-        return result
-
-    def flatten(self):
-        new = self.__class__()
-
-        # make sure we don't have duplicates
-        seen = {r.database.lower() for r in self if r.database}
-        if len(seen) > 1:
-            raise_compiler_error(str(seen))
-
-        for information_schema_name, schema in self.search():
-            path = {
-                'database': information_schema_name.database,
-                'schema': schema
-            }
-            new.add(information_schema_name.incorporate(
-                path=path,
-                quote_policy={'database': False},
-                include_policy={'database': False},
-            ))
-
-        return new
 
 
 class BaseAdapter(metaclass=AdapterMeta):
