+++ core/dbt/adapters/base/relation.py
@@ -6,7 +6,8 @@ import dbt.exceptions
 from collections.abc import Mapping, Hashable
 from dataclasses import dataclass, fields
 from typing import (
-    Optional, TypeVar, Generic, Any, Type, Dict, Union, List, Iterator, Tuple
+    Optional, TypeVar, Generic, Any, Type, Dict, Union, List, Iterator, Tuple,
+    Set
 )
 from typing_extensions import Protocol
 
@@ -496,3 +497,59 @@ class InformationSchema(BaseRelation):
         for k, v in super()._render_iterator():
             yield k, v
         yield None, self.information_schema_view
+
+
+class SchemaSearchMap(Dict[InformationSchema, Set[Optional[str]]]):
+    """A utility class to keep track of what information_schema tables to
+    search for what schemas
+    """
+    def add(self, relation: BaseRelation, preserve_case=False):
+        key = relation.information_schema_only()
+        if key not in self:
+            self[key] = set()
+        schema: Optional[str] = None
+        if relation.schema is not None:
+            if preserve_case:
+                schema = relation.schema
+            else:
+                schema = relation.schema.lower()
+        self[key].add(schema)
+
+    def search(self) -> Iterator[Tuple[InformationSchema, Optional[str]]]:
+        for information_schema_name, schemas in self.items():
+            for schema in schemas:
+                yield information_schema_name, schema
+
+    def schemas_searched(self) -> Set[Tuple[str, Optional[str]]]:
+        result: Set[Tuple[str, Optional[str]]] = set()
+        for information_schema_name, schemas in self.items():
+            if information_schema_name.database is None:
+                raise InternalException(
+                    'Got a None database in an information schema!'
+                )
+            result.update(
+                (information_schema_name.database, schema)
+                for schema in schemas
+            )
+        return result
+
+    def flatten(self):
+        new = self.__class__()
+
+        # make sure we don't have duplicates
+        seen = {r.database.lower() for r in self if r.database}
+        if len(seen) > 1:
+            dbt.exceptions.raise_compiler_error(str(seen))
+
+        for information_schema_name, schema in self.search():
+            path = {
+                'database': information_schema_name.database,
+                'schema': schema
+            }
+            new.add(information_schema_name.incorporate(
+                path=path,
+                quote_policy={'database': False},
+                include_policy={'database': False},
+            ))
+
+        return new
